Add files via upload
The code is for mapping OTP generated by the bank using ECC. The map maps the digits 0-9 to random numbers in the array eg: 0->(1,2). secp256k1 curve is chosen for mapping. The public key is generated from two random numbers Gx and Gy(points on the curve). The OTP is taken as an input. Private key is predefined which is a large number. The OTP is then encrypted using EccMultiply function. The message will be then decyphired at the other end using the public key and their private key.